Subhypercubes of magic hypercubes

Extracts all subhypercubes from an n-dimensional hypercube.

Usage

allsubhypercubes(a)

Arguments

a

The magic hypercube whose subhypercubes are computed

Value

Returns a list, each element of which is a subhypercube of a. Note that major diagonals are also returned (as n-by-1 arrays).

The names of the list are the extracted subhypercubes. Consider a <- magichypercube.4n(1,d=4) (so n=4) and if jj <- allsubhypercubes(a), consider jj[9]. The name of jj[9] is "n-i+1,i,i,"; its value is a square matrix. The columns of jj[9] may be recovered by a[n-i+1,i,i,] with \(i=1\ldots n\) (NB: that is, jj[[9]] == cbind(a[n-1+1,1,1,], a[n-2+1,2,2,], a[n-3+1,3,3,], a[n-4+1,4,4,]) where n=4).

The list does not include the whole array.

Author

Robin K. S. Hankin

Note

This function is a dog's dinner. It's complicated, convoluted, and needs an absurd use of the eval(parse(text=...)) construction. Basically it sucks big time.

BUT... I cannot for the life of me see a better way that gives the same results, without loops, on hypercubes of arbitrary dimension.

On my 256MB Linuxbox, allsubhypercubes() cannot cope with d as high as 5, for n=4. Heigh ho.

The term “subhypercube” does not include diagonally oriented entities at is.magichypercube. But it does here.

See also

Examples

  a <- magichypercube.4n(1,d=4)
  allsubhypercubes(a)
#> $`i,i,,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  240   81  145   48
#> [3,]  224   97  161   32
#> [4,]   49  144   80  241
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    6  187  123  198
#> [2,]  235   86  150   43
#> [3,]  219  102  166   27
#> [4,]   54  139   75  246
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   11  182  118  203
#> [2,]  230   91  155   38
#> [3,]  214  107  171   22
#> [4,]   59  134   70  251
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   16  177  113  208
#> [2,]  225   96  160   33
#> [3,]  209  112  176   17
#> [4,]   64  129   65  256
#> 
#> 
#> $`n-i+1,i,,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  189  125  196
#> [2,]  237   84  148   45
#> [3,]  221  100  164   29
#> [4,]   52  141   77  244
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    7  186  122  199
#> [2,]  234   87  151   42
#> [3,]  218  103  167   26
#> [4,]   55  138   74  247
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   10  183  119  202
#> [2,]  231   90  154   39
#> [3,]  215  106  170   23
#> [4,]   58  135   71  250
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  180  116  205
#> [2,]  228   93  157   36
#> [3,]  212  109  173   20
#> [4,]   61  132   68  253
#> 
#> 
#> $`i,n-i+1,,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  180  116  205
#> [2,]  228   93  157   36
#> [3,]  212  109  173   20
#> [4,]   61  132   68  253
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   10  183  119  202
#> [2,]  231   90  154   39
#> [3,]  215  106  170   23
#> [4,]   58  135   71  250
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    7  186  122  199
#> [2,]  234   87  151   42
#> [3,]  218  103  167   26
#> [4,]   55  138   74  247
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  189  125  196
#> [2,]  237   84  148   45
#> [3,]  221  100  164   29
#> [4,]   52  141   77  244
#> 
#> 
#> $`n-i+1,n-i+1,,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   16  177  113  208
#> [2,]  225   96  160   33
#> [3,]  209  112  176   17
#> [4,]   64  129   65  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   11  182  118  203
#> [2,]  230   91  155   38
#> [3,]  214  107  171   22
#> [4,]   59  134   70  251
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    6  187  123  198
#> [2,]  235   86  150   43
#> [3,]  219  102  166   27
#> [4,]   54  139   75  246
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  240   81  145   48
#> [3,]  224   97  161   32
#> [4,]   49  144   80  241
#> 
#> 
#> $`i,,i,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  252   69  133   60
#> [3,]  248   73  137   56
#> [4,]   13  180  116  205
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   18  175  111  210
#> [2,]  235   86  150   43
#> [3,]  231   90  154   39
#> [4,]   30  163   99  222
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   35  158   94  227
#> [2,]  218  103  167   26
#> [3,]  214  107  171   22
#> [4,]   47  146   82  239
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   52  141   77  244
#> [2,]  201  120  184    9
#> [3,]  197  124  188    5
#> [4,]   64  129   65  256
#> 
#> 
#> $`n-i+1,,i,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  189  125  196
#> [2,]  249   72  136   57
#> [3,]  245   76  140   53
#> [4,]   16  177  113  208
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   19  174  110  211
#> [2,]  234   87  151   42
#> [3,]  230   91  155   38
#> [4,]   31  162   98  223
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   34  159   95  226
#> [2,]  219  102  166   27
#> [3,]  215  106  170   23
#> [4,]   46  147   83  238
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  144   80  241
#> [2,]  204  117  181   12
#> [3,]  200  121  185    8
#> [4,]   61  132   68  253
#> 
#> 
#> $`,i,i,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  255   66  130   63
#> [3,]  254   67  131   62
#> [4,]    4  189  125  196
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   21  172  108  213
#> [2,]  235   86  150   43
#> [3,]  234   87  151   42
#> [4,]   24  169  105  216
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   41  152   88  233
#> [2,]  215  106  170   23
#> [3,]  214  107  171   22
#> [4,]   44  149   85  236
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   61  132   68  253
#> [2,]  195  126  190    3
#> [3,]  194  127  191    2
#> [4,]   64  129   65  256
#> 
#> 
#> $`i,i,i,`
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  235  214   64
#> [2,]  192   86  107  129
#> [3,]  128  150  171   65
#> [4,]  193   43   22  256
#> 
#> $`n-i+1,i,i,`
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  234  215   61
#> [2,]  189   87  106  132
#> [3,]  125  151  170   68
#> [4,]  196   42   23  253
#> 
#> $`,n-i+1,i,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  180  116  205
#> [2,]  243   78  142   51
#> [3,]  242   79  143   50
#> [4,]   16  177  113  208
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   25  168  104  217
#> [2,]  231   90  154   39
#> [3,]  230   91  155   38
#> [4,]   28  165  101  220
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   37  156   92  229
#> [2,]  219  102  166   27
#> [3,]  218  103  167   26
#> [4,]   40  153   89  232
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  144   80  241
#> [2,]  207  114  178   15
#> [3,]  206  115  179   14
#> [4,]   52  141   77  244
#> 
#> 
#> $`i,n-i+1,i,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  231  218   52
#> [2,]  180   90  103  141
#> [3,]  116  154  167   77
#> [4,]  205   39   26  244
#> 
#> $`n-i+1,n-i+1,i,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   16  230  219   49
#> [2,]  177   91  102  144
#> [3,]  113  155  166   80
#> [4,]  208   38   27  241
#> 
#> $`i,,n-i+1,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  144   80  241
#> [2,]  204  117  181   12
#> [3,]  200  121  185    8
#> [4,]   61  132   68  253
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   34  159   95  226
#> [2,]  219  102  166   27
#> [3,]  215  106  170   23
#> [4,]   46  147   83  238
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   19  174  110  211
#> [2,]  234   87  151   42
#> [3,]  230   91  155   38
#> [4,]   31  162   98  223
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  189  125  196
#> [2,]  249   72  136   57
#> [3,]  245   76  140   53
#> [4,]   16  177  113  208
#> 
#> 
#> $`n-i+1,,n-i+1,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   52  141   77  244
#> [2,]  201  120  184    9
#> [3,]  197  124  188    5
#> [4,]   64  129   65  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   35  158   94  227
#> [2,]  218  103  167   26
#> [3,]  214  107  171   22
#> [4,]   47  146   82  239
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   18  175  111  210
#> [2,]  235   86  150   43
#> [3,]  231   90  154   39
#> [4,]   30  163   99  222
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  252   69  133   60
#> [3,]  248   73  137   56
#> [4,]   13  180  116  205
#> 
#> 
#> $`,i,n-i+1,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  144   80  241
#> [2,]  207  114  178   15
#> [3,]  206  115  179   14
#> [4,]   52  141   77  244
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   37  156   92  229
#> [2,]  219  102  166   27
#> [3,]  218  103  167   26
#> [4,]   40  153   89  232
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   25  168  104  217
#> [2,]  231   90  154   39
#> [3,]  230   91  155   38
#> [4,]   28  165  101  220
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  180  116  205
#> [2,]  243   78  142   51
#> [3,]  242   79  143   50
#> [4,]   16  177  113  208
#> 
#> 
#> $`i,i,n-i+1,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  219  230   16
#> [2,]  144  102   91  177
#> [3,]   80  166  155  113
#> [4,]  241   27   38  208
#> 
#> $`n-i+1,i,n-i+1,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   52  218  231   13
#> [2,]  141  103   90  180
#> [3,]   77  167  154  116
#> [4,]  244   26   39  205
#> 
#> $`,n-i+1,n-i+1,`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   61  132   68  253
#> [2,]  195  126  190    3
#> [3,]  194  127  191    2
#> [4,]   64  129   65  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   41  152   88  233
#> [2,]  215  106  170   23
#> [3,]  214  107  171   22
#> [4,]   44  149   85  236
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   21  172  108  213
#> [2,]  235   86  150   43
#> [3,]  234   87  151   42
#> [4,]   24  169  105  216
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  192  128  193
#> [2,]  255   66  130   63
#> [3,]  254   67  131   62
#> [4,]    4  189  125  196
#> 
#> 
#> $`i,n-i+1,n-i+1,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   61  215  234    4
#> [2,]  132  106   87  189
#> [3,]   68  170  151  125
#> [4,]  253   23   42  196
#> 
#> $`n-i+1,n-i+1,n-i+1,`
#>      [,1] [,2] [,3] [,4]
#> [1,]   64  214  235    1
#> [2,]  129  107   86  192
#> [3,]   65  171  150  128
#> [4,]  256   22   43  193
#> 
#> $`i,,,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  240  224   49
#> [2,]  252   21   37  204
#> [3,]  248   25   41  200
#> [4,]   13  228  212   61
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   66  175  159  114
#> [2,]  187   86  102  139
#> [3,]  183   90  106  135
#> [4,]   78  163  147  126
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  131  110   94  179
#> [2,]  122  151  167   74
#> [3,]  118  155  171   70
#> [4,]  143   98   82  191
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  196   45   29  244
#> [2,]   57  216  232    9
#> [3,]   53  220  236    5
#> [4,]  208   33   17  256
#> 
#> 
#> $`n-i+1,,,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  237  221   52
#> [2,]  249   24   40  201
#> [3,]  245   28   44  197
#> [4,]   16  225  209   64
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   67  174  158  115
#> [2,]  186   87  103  138
#> [3,]  182   91  107  134
#> [4,]   79  162  146  127
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  130  111   95  178
#> [2,]  123  150  166   75
#> [3,]  119  154  170   71
#> [4,]  142   99   83  190
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   48   32  241
#> [2,]   60  213  229   12
#> [3,]   56  217  233    8
#> [4,]  205   36   20  253
#> 
#> 
#> $`,i,,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  240  224   49
#> [2,]  255   18   34  207
#> [3,]  254   19   35  206
#> [4,]    4  237  221   52
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   69  172  156  117
#> [2,]  187   86  102  139
#> [3,]  186   87  103  138
#> [4,]   72  169  153  120
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  137  104   88  185
#> [2,]  119  154  170   71
#> [3,]  118  155  171   70
#> [4,]  140  101   85  188
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  205   36   20  253
#> [2,]   51  222  238    3
#> [3,]   50  223  239    2
#> [4,]  208   33   17  256
#> 
#> 
#> $`i,i,,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  187  118  208
#> [2,]  240   86  155   33
#> [3,]  224  102  171   17
#> [4,]   49  139   70  256
#> 
#> $`n-i+1,i,,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  186  119  205
#> [2,]  237   87  154   36
#> [3,]  221  103  170   20
#> [4,]   52  138   71  253
#> 
#> $`,n-i+1,,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  228  212   61
#> [2,]  243   30   46  195
#> [3,]  242   31   47  194
#> [4,]   16  225  209   64
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   73  168  152  121
#> [2,]  183   90  106  135
#> [3,]  182   91  107  134
#> [4,]   76  165  149  124
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  133  108   92  181
#> [2,]  123  150  166   75
#> [3,]  122  151  167   74
#> [4,]  136  105   89  184
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   48   32  241
#> [2,]   63  210  226   15
#> [3,]   62  211  227   14
#> [4,]  196   45   29  244
#> 
#> 
#> $`i,n-i+1,,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  183  122  196
#> [2,]  228   90  151   45
#> [3,]  212  106  167   29
#> [4,]   61  135   74  244
#> 
#> $`n-i+1,n-i+1,,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   16  182  123  193
#> [2,]  225   91  150   48
#> [3,]  209  107  166   32
#> [4,]   64  134   75  241
#> 
#> $`,,i,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  252  248   13
#> [2,]  255    6   10  243
#> [3,]  254    7   11  242
#> [4,]    4  249  245   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   81  172  168   93
#> [2,]  175   86   90  163
#> [3,]  174   87   91  162
#> [4,]   84  169  165   96
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  161   92   88  173
#> [2,]   95  166  170   83
#> [3,]   94  167  171   82
#> [4,]  164   89   85  176
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  241   12    8  253
#> [2,]   15  246  250    3
#> [3,]   14  247  251    2
#> [4,]  244    9    5  256
#> 
#> 
#> $`i,,i,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  175   94  244
#> [2,]  252   86  167    9
#> [3,]  248   90  171    5
#> [4,]   13  163   82  256
#> 
#> $`n-i+1,,i,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  174   95  241
#> [2,]  249   87  166   12
#> [3,]  245   91  170    8
#> [4,]   16  162   83  253
#> 
#> $`,i,i,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  172   88  253
#> [2,]  255   86  170    3
#> [3,]  254   87  171    2
#> [4,]    4  169   85  256
#> 
#> $`i,i,i,i`
#> [1]   1  86 171 256
#> 
#> $`n-i+1,i,i,i`
#> [1]   4  87 170 253
#> 
#> $`,n-i+1,i,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  168   92  241
#> [2,]  243   90  166   15
#> [3,]  242   91  167   14
#> [4,]   16  165   89  244
#> 
#> $`i,n-i+1,i,i`
#> [1]  13  90 167 244
#> 
#> $`n-i+1,n-i+1,i,i`
#> [1]  16  91 166 241
#> 
#> $`,,n-i+1,i`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  204  200   61
#> [2,]  207   54   58  195
#> [3,]  206   55   59  194
#> [4,]   52  201  197   64
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   97  156  152  109
#> [2,]  159  102  106  147
#> [3,]  158  103  107  146
#> [4,]  100  153  149  112
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  145  108  104  157
#> [2,]  111  150  154   99
#> [3,]  110  151  155   98
#> [4,]  148  105  101  160
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   60   56  205
#> [2,]   63  198  202   51
#> [3,]   62  199  203   50
#> [4,]  196   57   53  208
#> 
#> 
#> $`i,,n-i+1,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  159  110  196
#> [2,]  204  102  151   57
#> [3,]  200  106  155   53
#> [4,]   61  147   98  208
#> 
#> $`n-i+1,,n-i+1,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   52  158  111  193
#> [2,]  201  103  150   60
#> [3,]  197  107  154   56
#> [4,]   64  146   99  205
#> 
#> $`,i,n-i+1,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  156  104  205
#> [2,]  207  102  154   51
#> [3,]  206  103  155   50
#> [4,]   52  153  101  208
#> 
#> $`i,i,n-i+1,i`
#> [1]  49 102 155 208
#> 
#> $`n-i+1,i,n-i+1,i`
#> [1]  52 103 154 205
#> 
#> $`,n-i+1,n-i+1,i`
#>      [,1] [,2] [,3] [,4]
#> [1,]   61  152  108  193
#> [2,]  195  106  150   63
#> [3,]  194  107  151   62
#> [4,]   64  149  105  196
#> 
#> $`i,n-i+1,n-i+1,i`
#> [1]  61 106 151 196
#> 
#> $`n-i+1,n-i+1,n-i+1,i`
#> [1]  64 107 150 193
#> 
#> $`i,,,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   48   32  241
#> [2,]   60  213  229   12
#> [3,]   56  217  233    8
#> [4,]  205   36   20  253
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  130  111   95  178
#> [2,]  123  150  166   75
#> [3,]  119  154  170   71
#> [4,]  142   99   83  190
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   67  174  158  115
#> [2,]  186   87  103  138
#> [3,]  182   91  107  134
#> [4,]   79  162  146  127
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4  237  221   52
#> [2,]  249   24   40  201
#> [3,]  245   28   44  197
#> [4,]   16  225  209   64
#> 
#> 
#> $`n-i+1,,,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  196   45   29  244
#> [2,]   57  216  232    9
#> [3,]   53  220  236    5
#> [4,]  208   33   17  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  131  110   94  179
#> [2,]  122  151  167   74
#> [3,]  118  155  171   70
#> [4,]  143   98   82  191
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   66  175  159  114
#> [2,]  187   86  102  139
#> [3,]  183   90  106  135
#> [4,]   78  163  147  126
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  240  224   49
#> [2,]  252   21   37  204
#> [3,]  248   25   41  200
#> [4,]   13  228  212   61
#> 
#> 
#> $`,i,,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   48   32  241
#> [2,]   63  210  226   15
#> [3,]   62  211  227   14
#> [4,]  196   45   29  244
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  133  108   92  181
#> [2,]  123  150  166   75
#> [3,]  122  151  167   74
#> [4,]  136  105   89  184
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   73  168  152  121
#> [2,]  183   90  106  135
#> [3,]  182   91  107  134
#> [4,]   76  165  149  124
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   13  228  212   61
#> [2,]  243   30   46  195
#> [3,]  242   31   47  194
#> [4,]   16  225  209   64
#> 
#> 
#> $`i,i,,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  193  123  182   16
#> [2,]   48  150   91  225
#> [3,]   32  166  107  209
#> [4,]  241   75  134   64
#> 
#> $`n-i+1,i,,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  196  122  183   13
#> [2,]   45  151   90  228
#> [3,]   29  167  106  212
#> [4,]  244   74  135   61
#> 
#> $`,n-i+1,,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  205   36   20  253
#> [2,]   51  222  238    3
#> [3,]   50  223  239    2
#> [4,]  208   33   17  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  137  104   88  185
#> [2,]  119  154  170   71
#> [3,]  118  155  171   70
#> [4,]  140  101   85  188
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   69  172  156  117
#> [2,]  187   86  102  139
#> [3,]  186   87  103  138
#> [4,]   72  169  153  120
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  240  224   49
#> [2,]  255   18   34  207
#> [3,]  254   19   35  206
#> [4,]    4  237  221   52
#> 
#> 
#> $`i,n-i+1,,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  205  119  186    4
#> [2,]   36  154   87  237
#> [3,]   20  170  103  221
#> [4,]  253   71  138   52
#> 
#> $`n-i+1,n-i+1,,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  208  118  187    1
#> [2,]   33  155   86  240
#> [3,]   17  171  102  224
#> [4,]  256   70  139   49
#> 
#> $`,,i,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  193   60   56  205
#> [2,]   63  198  202   51
#> [3,]   62  199  203   50
#> [4,]  196   57   53  208
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  145  108  104  157
#> [2,]  111  150  154   99
#> [3,]  110  151  155   98
#> [4,]  148  105  101  160
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   97  156  152  109
#> [2,]  159  102  106  147
#> [3,]  158  103  107  146
#> [4,]  100  153  149  112
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   49  204  200   61
#> [2,]  207   54   58  195
#> [3,]  206   55   59  194
#> [4,]   52  201  197   64
#> 
#> 
#> $`i,,i,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  193  111  158   52
#> [2,]   60  150  103  201
#> [3,]   56  154  107  197
#> [4,]  205   99  146   64
#> 
#> $`n-i+1,,i,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  196  110  159   49
#> [2,]   57  151  102  204
#> [3,]   53  155  106  200
#> [4,]  208   98  147   61
#> 
#> $`,i,i,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  193  108  152   61
#> [2,]   63  150  106  195
#> [3,]   62  151  107  194
#> [4,]  196  105  149   64
#> 
#> $`i,i,i,n-i+1`
#> [1] 193 150 107  64
#> 
#> $`n-i+1,i,i,n-i+1`
#> [1] 196 151 106  61
#> 
#> $`,n-i+1,i,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  205  104  156   49
#> [2,]   51  154  102  207
#> [3,]   50  155  103  206
#> [4,]  208  101  153   52
#> 
#> $`i,n-i+1,i,n-i+1`
#> [1] 205 154 103  52
#> 
#> $`n-i+1,n-i+1,i,n-i+1`
#> [1] 208 155 102  49
#> 
#> $`,,n-i+1,n-i+1`
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  241   12    8  253
#> [2,]   15  246  250    3
#> [3,]   14  247  251    2
#> [4,]  244    9    5  256
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]  161   92   88  173
#> [2,]   95  166  170   83
#> [3,]   94  167  171   82
#> [4,]  164   89   85  176
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   81  172  168   93
#> [2,]  175   86   90  163
#> [3,]  174   87   91  162
#> [4,]   84  169  165   96
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1  252  248   13
#> [2,]  255    6   10  243
#> [3,]  254    7   11  242
#> [4,]    4  249  245   16
#> 
#> 
#> $`i,,n-i+1,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  241   95  174    4
#> [2,]   12  166   87  249
#> [3,]    8  170   91  245
#> [4,]  253   83  162   16
#> 
#> $`n-i+1,,n-i+1,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  244   94  175    1
#> [2,]    9  167   86  252
#> [3,]    5  171   90  248
#> [4,]  256   82  163   13
#> 
#> $`,i,n-i+1,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  241   92  168   13
#> [2,]   15  166   90  243
#> [3,]   14  167   91  242
#> [4,]  244   89  165   16
#> 
#> $`i,i,n-i+1,n-i+1`
#> [1] 241 166  91  16
#> 
#> $`n-i+1,i,n-i+1,n-i+1`
#> [1] 244 167  90  13
#> 
#> $`,n-i+1,n-i+1,n-i+1`
#>      [,1] [,2] [,3] [,4]
#> [1,]  253   88  172    1
#> [2,]    3  170   86  255
#> [3,]    2  171   87  254
#> [4,]  256   85  169    4
#> 
#> $`i,n-i+1,n-i+1,n-i+1`
#> [1] 253 170  87   4
#> 
#> $`n-i+1,n-i+1,n-i+1,n-i+1`
#> [1] 256 171  86   1
#>